rename controllers and get context
