+++ src/NuGetGallery/Areas/Admin/Views/Home/Index.cshtml
@@ -96,7 +96,7 @@
         </li>
         <li>
             <h2>
-                <a class="action-menu-link" href="@Url.Action(actionName: "Index", controllerName: "ReserveNamespace")">
+                <a class="action-menu-link" href="@Url.Action(actionName: "Index", controllerName: "ReservedNamespace")">
                     <i class="icon-user action-menu-icon"></i>
                     <span class="action-menu-text">Reserve Namespace</span>
                 </a>
